Refs #17404 Added notice builder
